LayoutTests/fast/canvas/webgl/webgl2-texStorage-expected.txt

CONSOLE MESSAGE: line 1: TypeError: null is not an object (evaluating 'gl.NO_ERROR')
Test that texStorage2D() works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
